Plug in the AP RO verification implementation
This patch enables the invocation of the AP RO verification function
in response to the operator entering the 'magic sequence' of holding
the power button pressed and pressing/releasing the refresh button
three times within five seconds.

The code used during the 'Open box RMA' verification process is used,
with the physical presence confirmation phase bypassed.

BUG=b:153764696, b:154966209
TEST=generated AP integrity verification data using the ap_ro_hash.py
     script and then ran the verification procedure, observing the
     'hash match' message on the Cr50 console.

     Also verified that the Open Box RMA procedure still succeeds.

+++ common/ap_ro_integrity_check.c
@@ -175,7 +175,7 @@ int validate_ap_ro(void)
 	uint8_t digest[SHA256_DIGEST_SIZE];
 	int rv;
 
-	if (p_chk->header.num_ranges == ~0) {
+	if (p_chk->header.num_ranges == (uint16_t) ~0) {
 		CPRINTS("%s: RO verification not programmed", __func__);
 		return EC_ERROR_INVAL;
 	}
@@ -239,7 +239,7 @@ static int ap_ro_info_cmd(int argc, char **argv)
 		flash_physical_erase(h1_flash_offset_, AP_RO_DATA_SPACE_SIZE);
 	}
 #endif
-	if ((p_chk->header.num_ranges == ~0) &&
+	if ((p_chk->header.num_ranges == (uint16_t) ~0) &&
 	    (p_chk->header.checksum == ~0)) {
 		ccprintf("AP RO check space is not programmed\n");
 		return EC_SUCCESS;
